feat: delete note

+++ src/repositories/deleteCredential.ts
@@ -1,10 +1,20 @@
 import { prisma } from "../config/database";
 
-export async function deleteCredential(id : number) {
-    await prisma.credentials.delete({
-        where: {
-            id
-        }
-    })
-    return;
+export async function deleteCredential(id : number, table : "credentials" | "notes") {
+    if(table === "credentials"){
+        await prisma.credentials.delete({
+            where: {
+                id
+            }
+        })
+        return;
+    }
+    if(table === "notes"){
+        await prisma.notes.delete({
+            where: {
+                id
+            }
+        })
+        return;
+    }    
 };
